Pktgen dpdk 350 upgrade



See merge request !4

+++ app/pktgen-port-cfg.h
@@ -162,7 +162,7 @@ typedef struct port_info_s {
 	uint32_t mpls_entry;	/**< Set the port MPLS entry */
 	uint32_t gre_key;	/**< GRE key if used */
 
-	uint16_t nb_mbufs;	/**< Number of mbufs in the system */
+	uint32_t nb_mbufs;	/**< Number of mbufs in the system */
 	uint16_t pad1;
 	uint64_t max_latency;	/**< TX Latency sequence */
 	uint64_t avg_latency;	/**< Latency delta in clock ticks */
